# Conflicts:
#	app/build.gradle.kts
#	gradle/libs.versions.toml

+++ domain/vod/src/main/java/entity/movie.kt
@@ -2,11 +2,11 @@ package entity
 import kotlinx.datetime.LocalDate
 
 data class Movie(
-    val id: Long,
+    val id: Int,
     val posterImageUrl: String,
     val title: String,
     val description: String,
-    val actorIds: List<Long>,
+    val actorIds: List<Int>,
     val releaseDate: LocalDate,
     val genres: List<Genre>,
     val imdbRating: Float?
